requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-ninja==1.3.0
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
